requirements.txt
langchain>=0.1.0
langchain-aws>=0.1.0
langchain-community>=0.0.13
langgraph>=0.0.20
boto3>=1.28.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
